Implement PR review and split up ethereum queries
