+++ application/ledger_query_service/src/ethereum/queries/mod.rs
@@ -1,0 +1,7 @@
+pub mod block_query;
+pub mod event_query;
+pub mod transaction_query;
+
+pub use self::{
+    block_query::BlockQuery, event_query::EventQuery, transaction_query::TransactionQuery,
+};
